Update directions.py to allow uniformity selection
New functions: 
make_directions_file_from_skymodel_uniform
NU
chooseGroupSize

**(experimental users can substitute make_directions_file_from_skymodel for make_directions_file_from_skymodel_uniform)**

Explanation:
One can derive the L2 measure of non-uniformity of the autocorrelation of calibrator locations on the sky, and we can then try to minimize the non-uniformity of directions_max_num calibrators by searching the brightest directions_max_num + searchDepth candidates for the distribution that minimizes it. Turns out this requires time proportional to (groupSize + searchDepth)! / (searchDepth! * groupSize!), so we can break this down into iterative sub-solutions and hope that the sum of uniformly distributed sub-groups is also uniform. If you search for 30 facets, and break this into 2 groups of 15 then this can be done in a few minutes using a dozen cores. However, 15+15 may not result in maximal uniformity because it is an approximation. Passing make_directions_file_from_skymodel_uniform maxTime=(number of minutes != None will try to estimate decent groupSize and searchDepth).